Bug 526591 - fix "using hashCode() on array"
Calling hashCode() on an array returns the same value as
System.identityHashCode(), and ignores the contents of the array. If
this happens in hashCode() method of a class implementing equals() and
the class uses the array content for equals(), it will break the
contract of the hashCode() method saying that equal objects should
return same hash value.

Change-Id: I0a986c0e5ed6e441588540001efb4ada35b2dbbd
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>