Bug 537666 - fix test for not existing class folders on Windows
On Windows, external class folders are always encoded with a device part
in the path. Assume that if the device part is missing on Windows, the
path to check can't represent external folder. Of course this could be a
path encoded on Linux, but in this case the project classpath is not
portable anyway.

Change-Id: I3fbd8546809176e33d4d9828310b0980a4b4e70a
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>