Bug 544921 - [performance] Poor performance with large source files
This patch uses sorted arrays in the CompilationUnitScope to improve
the runtime behavior with large source files where a large number of
dependent types exist and need to be recorded in the
qualifiedReferences, simpleNameReferences and rootReferences.

Lookup of memberTypes by name from a ReferenceBinding does use binary
search logic, too (similar approach as with fields).

Sorting of char[][] is extracted to SortedCharArrays.

CompoundNameVector is removed and replaced by SortedCompoundNameVector.
SimpleNameVector is removed and replaced by SortedSimpleNameVector.

Change-Id: Ib66c7f0399cdea0c5558a15bf835406ceb828988
Signed-off-by: Sebastian Zarnekow <sebastian.zarnekow@gmail.com>
Also-By: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>