Bug 549457 - Autobuild does nothing on org.eclipse.jdt.core.prefs change
If the file .settings/org.eclipse.jdt.core.prefs changes, JDT does not
automatically rebuild the project. E.g. pulling different settings from
git will not result in a re-build; only changing the settings from the
Java compiler preference page will ask the user if they wish to rebuild
the workspace.

This change adjusts JavaBuilder to inspect deltas for changes in
.settings/org.eclipse.jdt.core.prefs. In case of such, a full build is
done. It also adds a test to ensure build participants are triggered on
changes in .settings/org.eclipse.jdt.core.prefs changes.

In addition to this change, its necessary to adapt the dialog in the
Java compiler preference page (in JDT UI), to not give the option "no
build on preferences apply". This is the case, since applying the
compiler preferences will cause the auto-build to do a full build,
regardless of whether the user chose the "no" option.

In case this new behavior causes problems, e.g. in tooling which
re-generates project files without actually changing them, it is
possible to return to the previous behavior with the VM property:

-Dorg.eclipse.jdt.core.disableAutoBuildOnSettingsChange=true

Change-Id: I10fb9bb828eec43fcdd81d36c78f3e30aa1cc9dc
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>