Bug 481323 - support substring and subword matching in findTypes
Substring matching is always enabled and can only be disabled by the
existing VM arg -Djdt.codeCompleteSubstringMatch=false.
Subword matching is enabled by default and can be disabled with the
existing preference.

Modified SearchableEnvironment, SearchPattern, Index and
BasicSearchEngine to support match rules with any combination of exact
type, camel case, substring and subword matches for types.

In Substring/SubwordCompletionTests all java.* type proposals are
ignored. The intention is to avoid test maintenance when new types with
similar names are added to the JDK (e.g. CyclicBarrier is now a match
for "bar").

Change-Id: Ie51b96f0b20025001b7babcaf5aff1637bc0922e
Signed-off-by: Julian Honnen <julian.honnen@vector.com>